requirements.txt
Django==1.6
Pillow==2.7.0
decorator==3.4.0
django-tinymce==1.5.3
pytz==2014.4
six==1.7.3
uWSGI==2.0.10
wsgiref==0.1.2
